expression: split multi valued index test
