+++ expression/BUILD.bazel
@@ -183,7 +183,6 @@ go_test(
         "grouping_sets_test.go",
         "helper_test.go",
         "main_test.go",
-        "multi_valued_index_test.go",
         "scalar_function_test.go",
         "schema_test.go",
         "typeinfer_test.go",
@@ -211,7 +210,6 @@ go_test(
         "//sessionctx/stmtctx",
         "//sessionctx/variable",
         "//sessiontxn",
-        "//table",
         "//tablecodec",
         "//testkit",
         "//testkit/testmain",
