+++ expression/test/multivaluedindex/multi_valued_index_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package expression_test
+package multivaluedindex
 
 import (
 	"context"
